Update LendingPosition import in types
