+++ packages/dma-library/src/types/aave-like/aave-like-position-v2.ts
@@ -2,7 +2,7 @@ import { Address } from '@deploy-configurations/types/address'
 import { ZERO } from '@dma-common/constants'
 import { negativeToZero, normalizeValue } from '@dma-common/utils/common'
 import { LendingCumulativesData } from '@dma-library/types'
-import { LendingPosition } from '@dma-library/types/morphoblue/morphoblue-position'
+import { LendingPosition } from '@dma-library/types/common'
 import { ReserveData } from '@dma-library/views/aave/types'
 import { getBuyingPower } from '@dma-library/views/common'
 import { IPositionCategory, RiskRatio } from '@domain'
