+++ packages/dma-library/src/types/morphoblue/morphoblue-position.ts
@@ -1,43 +1,11 @@
 import { Address } from '@deploy-configurations/types/address'
 import { ONE, ZERO } from '@dma-common/constants'
 import { negativeToZero, normalizeValue } from '@dma-common/utils/common'
+import { LendingPosition } from '@dma-library/types/common'
 import { getBuyingPower } from '@dma-library/views/common'
 import { MorphoCumulativesData } from '@dma-library/views/morpho'
-import { IRiskRatio, RiskRatio } from '@domain'
+import { RiskRatio } from '@domain'
 import { BigNumber } from 'bignumber.js'
-
-export interface LendingPosition {
-  owner: Address
-  collateralAmount: BigNumber
-  debtAmount: BigNumber
-
-  marketPrice: BigNumber
-  liquidationPrice: BigNumber
-  liquidationToMarketPrice: BigNumber
-
-  collateralAvailable: BigNumber
-  riskRatio: IRiskRatio
-  maxRiskRatio: IRiskRatio
-  minRiskRatio: IRiskRatio
-
-  borrowRate: BigNumber
-  netValue: BigNumber
-  buyingPower: BigNumber
-  pnl: {
-    withFees: BigNumber
-    withoutFees: BigNumber
-  }
-
-  debtAvailable(collateralAmount?: BigNumber, debtAmount?: BigNumber): BigNumber
-
-  deposit(amount: BigNumber): LendingPosition
-
-  withdraw(amount: BigNumber): LendingPosition
-
-  borrow(amount: BigNumber): LendingPosition
-
-  payback(amount: BigNumber): LendingPosition
-}
 
 interface MarketParams {
   id: string
